Wris_flood_model/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1
